Bug fix found by Vincent Meunier.  The wannier centers were calculated incorrectly, and are now fixed....EJB
correspondence is as follows:

Kiril, Eric,

I do really appreciate your support with using the wannier function calculator in nwchem. You've sent to me a large number of precious tips and information that will surely help me getting what I need to realize my original ideas.

Something in what you told me kind of worried me from the beginning, namely the fact that the WF centers were only known "modulo half a lattice constant". This is surprising since, for a large enough cutoff, the WFs are well converged and the center of them seems to be unambiguous, at least after plotting them. I've tried to find out why there is such an ambiguity and here are my findings. First and foremost, I'm ready to say I'm wrong if you convince me my arguments are erroneous.

I believe the reason why there is an ambiguity comes from the fact that the procedure that is coded makes use of the intrinsic fortran function atan while it should actually be using atan2. [as you know atan[x] returns the inverse tangent of x in the range [-pi/2,pi/2], and when divided by the lattice constant in reciprocal space, one sees that indeed one can not span the all unit cell and one gets a wrong (not only unambiguous) center. On the contrary, atan2[x,y] will allow for getting the right answer since it returns a value in the full range [-pi,pi], in wich case the uncertainty is simply an integer number of lattice constants as it should be for a periodic system]  I convinced myself this was true by using the alternate definition from Silverstelli paper, who makes use of -Im[log[X]], which for X=AExp[-iB], will return B +2 n Pi (n integer) as it should be, while Atan[SIN[B]/COS[B]] returns B + n Pi.

I've tested that on a simple system: the carbon chain. The details of the results are summarized in the attached pdf file.

Do you agree with me?

Thanks,

Vincent