HvD: Added the option to set the scratch and permanent directories, as well as the memory limits in the nwchemrc files or with environment variables. Examples of this are included in gen_default.nwchemrc. At present these examples are commented out but this way there is some minimalistic documentation of this capability.
In general terms I still have a question mark over how the scratch
directory is handled. At present the setting in the environment or 
nwchemrc overrides the value stored in the RTDB in restart calculations.
(If the user specifies scratch_dir in the input then that takes 
precedence anyway, so most users will probably never notice this). 
It is likely that in case of a restart we want the value from the RTDB
to take precedence over the environment settings. I hope experience
will generate insight in how exactly to handle this situation. Of course
developers smarter than me are welcome to comment.