HvD: Previously the geom_rtdb_load function could not distinguish been corrupted RTDB's and geometries simply not having been stored. In all cases warning messages would be generated on standard output. This made it impossible to just check whether a geometry is present on the RTDB or not without sending scary messages to the user.
This checkin fixes this in the following way. If not even the first
item of a geometry is present on RTDB then the geometry is assumed 
absent and the routine returns .false. but remains silent otherwise.
If the first item of a geometry is present, then all expected items
must be there otherwise the data on the RTDB is clearly corrupt. So
if some data is not there, return .false. and print a warning on 
standard output.

This fix preserves sending warnings when something is clearly wrong,
but also allows one to test for the presence of a geometry without
generating terrifying messages when benign exceptional conditions arise.