HvD: Doxygen ran into trouble because it could not find most of the include files NWChem uses. This caused segmentation faults, and parts of the documentation getting rejected. Adding the directories that hold include files to the Doxygen configuration seems to improve this situation significantly.
Adding the include file directories is however not trivial. Doxygen assumes
that these directories can essentially be hardwired in the configuration file.
This would mean that your source code always has to live in the same directory,
and you have to run Doxygen always in the same directory. This lack of
flexibility seems annoying. The alternative is to configure Doxygen every time
you run it. This can be done relatively easily, but it adds about 2 minutes to
the execution time. Also at the moment we assume that you are running the 
run_doxygen script that is part of the repository on which you are running it.
The assumption stems from the fact that the location of the source code
directory is specified relative to the contrib/doxygen directory.