HvD: Extended the ga_lkain and ga_kain linear systems solvers to handle complex valued problems in addition to the double precision systems it could do before. The API remains the same as before. The only difference is that for the complex case the global arrays g_x and g_b have to be created with the MT_DCPL type. In addition the routines for the matrix-vector products and the preconditioner that are passed to these routines have to be implemented for the complex type as well. The shift passed to the preconditioner is still of the double precision type.
As an aside util_flush now also calls util_cflush. This change is related to
the way that the GNU compilers buffer output. They buffer Fortran standard out
and C standard separately which can lead to confusing and/or messy mixing in 
the output. Therefore it is better to flush both Fortran and C I/O. If someone
knows a way to eliminate this call for none GNU compilers, please let me know.