HvD: A better fix for the solvation module stubbing problem. The issue was that when the solvation code was part of the property module it would get stubbed out with the property module. Now the solvation code is a module of its own, and there it must now be possible to stub it independently.
This is particularly true in the case of using the hessian module. Building
the hessian module requires the property module, but we do not have solvation
hessian capabilities. So the hessian module should not be required to pull in 
the solvation module in addition to the property module. However, this requires
the solvation module to be stubbed, whereas the property module is required to
be built. Hence the code can only be built this way if the solvation routines
can be stubbed separately.