HvD: In saddle point optimizations where some of the eigenvalues of the Hessian approach 0 the code tried to do clever things. In particular if the gradient along the 0 frequency mode was 0 also the code would set the step to 0. The problem was that the gradient was considered 0 only if it was smaller than the gradient convergence criterion divided by 100. Given that the gradient convergence criterion was 1.5e-5 the gradient had to drop below 1.5e-7 for the step to be set to 0. Otherwise the code would continue to use a Newton step. Given that the eigenvalue of the Hessian in a user reported case was of the order of 1.0e-5 the result was that the step size never reached 0 and the code would never converge no matter how many optimization steps you ran.
My take on this is that if the Hessian has a 0 eigenvalue the Newton method
cannot be used. Instead you are better off using steepest descent. So simply
use steepest descent with a step size restriction, i.e.

   step = sign(min(trust,abs(grad)),-grad)

and do not try to be clever. This fixed the reported problem although it still
took 134 steps to converge.