HvD: An example of the insanity that results from generating code with a symbolic algebra package but without using proper optimization. This subroutine is 41 MB or 565,000 lines in size. The Fortran standard allows a single statement to take up to 40 lines, the Intel compiler extends this to 512 lines. Still the Intel compiler rejects this code with
./f77/nwxcm_c_m06.F(211023): error #5157: Too many continuation lines
./f77/nwxcm_c_m06.F(213690): catastrophic error: Statement too long

The GNU gfortran compiler happily attempts to compile this but essentially
runs out of memory during the optimization.

In a next incarnation of the autoxc script we will have Maxima optimize
the expressions and see what that will do. Still I would like to keep a copy
of this version of this file just as a deterrent.