HvD: Adding a first pass of the 2nd and 3rd derivatives generated with the Maxima symbolic algebra program. Starting the code generation on Sunday at 10:41 AM and finishing on Wednesday at 00:11 AM it took approximately 62 hours to generate this code (for most part using two cores on my machine). The main offender was the TPSS correlation functional which took well over 12 hours. Part of this was due to Maxima exhausting all 4 GB of physical memory in my desktop machine and the subsequent paging reduced the performance significantly.
Nevertheless the results look good right now (still need to test the higher
order derivative which could significantly impact this assessment). In 
particular the code for the M06 correlation functional was originally 44 MB
in size. This was the result of simply generating the relevant derivative
expressions and expressing them as Fortran code. Now that the autoxc script
has been enhanced to optimize the expressions before generating Fortran the
same routines only amount to 470 KB (a reduction of almost a factor 100 (98 to
be precise)).

It is obvious that optimizing these expressions does take time. However, the
good thing is that with this result the compiler does not need to take that time
when you try to build NWChem. I am sure that anyone compiling NWChem will be
grateful for not having to wait 62 hours.