HvD: Automatic differentiation can only compute derivatives for the arguments provided to a function. For DFT that means that in the closed shell case where only rho, gamma and tau are provided only derivatives wrt. to those quantities can be calculated. In particular this means that derivatives wrt. quantities such as rho_a, gamma_aa, and gamma_ab, etc. are not available.
If you are doing a closed shell calculation (the TDDFT triplet excited
states for a closed shell ground state molecule is a special case, the
densities of the alpha- and beta-spin channels are the same but the
triplet energy expression depends on the spin in a non-trivial way and
therefore you need the derivatives for the separate spin-components)
usually the required derivatives can be linked to the available
derivatives by a simple scale factor. I have worked these scale factors
out for the 1st and 2nd order derivatives. Hence one can now compute
TDDFT singlet excited states with the closed shell automatic
differentiation implementation of the density functionals.

The scale factors for the 3rd order derivatives still need doing. The
other thing that is still needed is a mechanism to switch the 
density functional evaluation to the unrestricted open shell case 
when doing a TDDFT triplet excited state. At the moment there is no
such mechanism in the code at all as the code assumes that you always
get the spin-components separated out. 

Comment: Interesting is that the symbolic algebra generated code could
in principle support three different modes:
1. Unrestricted open shell (alpha- and beta-densities are different
   and all spin-components are required)
2. Closed shell but spin-componentized (essentially evaluate the UHF
   expressions but exploit the fact that the alpha- and beta-densities
   are the same for some efficiency gain)
3. Closed shell (only use the total density and spin essentially does
   not exist).
The current symbolic algebra generated code supports modes 1 and 2,
but 3 is not implemented. (By contrast by its very nature the 
automatic differentiation code can only support modes 1 and 3).