HvD: Silly me, I had assumed that SVN would preserve the commit times of the source code files. This can actually be done by modifying your .subversion/config file and setting use-commit-times but we cannot expect our users to do that. In addition there would be knock on effects that are probably worse than the problem I am trying to fix.
Anyway as the timestamps on the f77/*.F files are randomly different from those
of the max/*.max files the NWChem build would willy-nilly try to regenerate
the Fortran source code through Maxima. This fails on most machines as Maxima
is not installed and otherwise would be highly undesirable as this code 
generation is hideously expensive (which is why the f77/*.F were checked-in
in the first place).

The solution is to suppress the dependency of the f77/*.F files on the max/*.max
files. A comment has been inserted to explain what is going on and what one
should do to add a new file to the f77/*.F collection. This little bit of
manual intervention should cause only a minor inconvience.