add more efficient getppn for BGQ and MPI-3
BGQ has a system call for PPN etc.
MPI-3 has a routine to get a node communicator and the size of this communicator is the number of PPN.

I have efficient implementations for Cray, but since Cray has supported MPI-3 for over a year, there is no need.

The MPI-2 implementation could be optimized but it should not be bottleneck and is unlikely to be used except
by users that insist on using old MPI implementations, since all relevant platforms support MPI-3 or are BGQ.