HvD: For some reason the atoms are now sorted in order of decreasing row in the periodic table. As a consistency check the code tests whether the final number of centers is the same as what went in. If the number does not match the code, sensibly, aborts.
The problem was that all first row atoms where considered those with a charge
less than or equal to 2 but larger than 0. This then excludes Bq and X centers,
which in the case of BSSE calculations are significant (see QA test cases such
dft_bsse and its variants). The BSSE test cases as a result crashed.

To fix this I have changed the definition of first row atoms to be all centers
with a charge less than or equal to 2 but larger than -1. This way Bq and X
centers are preserved and the BSSE test cases pass.