eliminate product reductions in ga_igop calls
'*' reduction maps to MPI_PROD, which is otherwise unused in
all of open-source parallel computing (to my knowledge).
since the reduction is just testing for the case where
every process has 1 (and not 0), we can instead use 'min'.

the best solution would be to use actual logical types
and logical reductions (e.g. MPI_LOGICAL and MPI_LAND)
but there is no way to access these from the current
GA collective interface.