replace quadruple loop around DAXPY with pure loops
particularly when BLAS uses threads, loops around BLAS calls can lead
to excessive and largely useless threading overheads.

any decent Fortran compiler will generate code that is equivalent to
a tuned DAXPY implementation because this kernel is bandwidth-limited
and thus barely benefits from SIMD and certainly doesn't benefit from
cache optimizations.

the benefit of the pure loop code is that it can be threaded at the
outermost loop, which minimizes overheads.

actually threaded the loops in question is future work.