move transcendentals inside of thresholded loop
transcendentals are expensive.  we can do the thresholding
without the exp() or sqrt() by using a few multiplications
and a single log().  this way, we avoid taking 3 exp() of
numbers that do not contribute.

we also used double loops and ditched the ipair structure.
we now do not write the end of ipair but that portion should
never be used since the whole point of the m2 loop is to
compress/sparsify it.

these changes pass both intchk and small_intchk.