TCE GPU: add environment varibale NWC_OFFLOAD_SPAN
The NWC_OFFLOAD_SPAN environment variable is meant to control which GA
ranks are offloading to GPUs.  When setting the variable, every
NWC_OFFLOAD_SPANth rank (staring at rank 0) will be offloading its work
to a GPU.

The current solution is a place holder for the final implementation that
still needs to be figured out with actual performance tests.