HvD: The ADIOS code needs to be called on the first and the last point
The ADIOS code needs some initialization on the first point and some
finalization on the last point. Hence the code needs to be called on
these two points. However, the step directive in the NWChem input
allows to run a number of points before data acquisition and to
skip a number of steps between data acquisition points. This could
mean that no ADIOS code was called on either the first or the last
step. This caused the code to fail.
Now we force a save on the first and the last point in addition
to any other points the code would normally have saved.