HvD: Check that GAs have enough space to hold the data for each processor.
Constructs like:

        call ga_distribution(ga_iw,node,ili,ihi,jli,jhi)
        ili=ili+nold
        ihi=ili+new-1
        call ga_put(ga_iw,ili,ihi,jli,jhi,iwl,mwm)

were causing problems as there is no guarantee that ihi in the put call
is smaller than the ihi from the distribution call. I have tracked all
such cases down and explicitly check that the GA is big enough to hold
the data. It seems that in all cases the GA size can be adjusted by
changing "msa" in the input file.

Symptoms of getting this wrong: some atoms disappear, and some atoms
show up on more than one processor. Usually the missing atoms are
detected when attempting to write the trajectory information.