HvD: add temperature capping to "vreass" directive.
If the keyword "max" is added to the "vreass" directive then
the temperature tends towards the target value but cannot
exceed it. It can be lower though. This way it is possible to
run a dynamics simulation where the chemical system tends towards
the lowest regions of the potential energy surface. I.e. you
can minimize the energy without running a conventional minimization
which might get stuck in a local minimum.