HvD: Fixing derivative generators
First of all I had to change the order of the indeces. This is because
in variadic templates the variadic type set has to come at the end.
For derivatives (which in principle can go to arbitrary order with
derivatives in respect to arbitrary types) this means that indeces that
correspond to variables that the derivatives are taken with respect to
have to come last in the tensors.
Also, I wasn't generating all the required derivatives.