HvD: The 1NT5 transmembrane protein
This case has been run both in solution as well as in vacuo.
This transmembrane protein consists of two different parts that are
normally bound through Hydrogen bonds. In these simulations the two
proteins are placed at 1, 2, 3, 4, and 5 nm separation. During the
simulation we hope to see them recombine.
In solution recombination happens for the proteins separated by
1, and 2 nm. If the proteins are separated by a larger distance
at the start then their motions are essentially independent and they
drift apart.
In vacuo the recombination happens rather quickly for every
separation. It is most obvious for proteins separated by 5 nm at
the beginning, mainly because the separation is really clear.
The structures labeled by separation are:
- 1nt5_mod1.pdb
- 1nt5_mod2.pdb
- 1nt5_mod3.pdb
- 1nt5_mod4.pdb
- 1nt5_mod5.pdb
The calculations in solution are prepared in files
- 1nt5_prep?.nw
The calculations in vacuo are prepared in files
- 1nt5_prep1?.nw