Libint orders spherical harmonic differently from NWChem.
In NWChem the spherical harmonic p-functions are ordered p_x, p_y
and p_z as they are the same as the Cartesian basis functions anyway.

In Libint spherical harmonic p-functions are ordered according to
the "angular momentum" (only the p_z is really an eigenfunction of the
angular momentum operator but let us not worry about that). So
Libint orders the p-functions p_y, p_z, and p_x.

NWChemEx also only uses spherical harmonic basis sets at least for
testing in Mokup.

So to compare integrals between NWChem and Libint for the development
of NWChemEx I need to reorder the p-functions before storing and
printing them.